refactor(app,ui): split overlay rendering, extract chip helper, clarify status-bar docs
Code-review follow-up to the entry-aware keymap fit (242fa4d):

- internal/app/view_overlays.go (new): house all renderOverlay* methods,
  convertNetpolRule, and clampErrorLogLines. view_status.go drops from
  925 to 451 lines, back under the 800-line guideline; the file is now
  focused on status-bar / breadcrumb / column-header concerns.
- internal/app/view_status.go: extract explorerStatusChips() from
  statusBar() (183 -> 149 lines). The chip-group composition (sort,
  counter / selection badge, filter preset, NYAN) and its rationale
  comment moved together so the why stays next to the code.
- internal/ui/hintbar.go: spell out the always-preserve-a-1-col-gutter
  behavior in JoinStatusBar's doc comment, and fix the leftBudget
  comment in view_status.go that incorrectly claimed "matches the
  elastic-spacer minimum" (the elastic minimum is 1; we reserve 2 for
  a comfortable visual gutter).

No behavior change; all tests pass with -race, golangci-lint clean,
coverage held (app 77.9 -> 78.0%, ui 85.4%).
